Expr.Class: remove negated field and parsing fix
remove the 'negated' field.  this property is now implicit when over 1/2
of the bitset's bits are set.  this shrinks Expr.Class from 40 bytes to
32.

parsing: previously peg-parsers.class parser was not special casing a
leading '^'.  now this is accounted for.  the 'class' parser will now
print an error message and return an error if more than 1/2 of the bits
are set.